Give assault defenders infinite respawns. Bump assault attacker respawns to 10. Decrease AI accuracy and aiming. Increase turret destruction reward. Switch to 64-bit secret server key. Make Library, Plaza, and Refinery more walkable for minions.

+++ src/settings.h
@@ -29,7 +29,7 @@ namespace Settings
 	extern s32 framerate_limit;
 	extern s32 display_mode_index;
 #if SERVER
-	extern s32 secret;
+	extern u64 secret;
 	extern u16 port;
 	extern char itch_api_key[MAX_AUTH_KEY + 1];
 	extern char public_ipv4[NET_MAX_ADDRESS];
